+++ app/Http/Controllers/Import/Nordigen/LinkController.php
@@ -31,6 +31,8 @@ use App\Http\Controllers\Controller;
 use App\Http\Middleware\LinkControllerMiddleware;
 use App\Services\Nordigen\Request\GetRequisitionRequest;
 use App\Services\Nordigen\Request\PostNewRequisitionRequest;
+use App\Services\Nordigen\Request\PostNewUserAgreement;
+use App\Services\Nordigen\Response\NewUserAgreementResponse;
 use App\Services\Nordigen\Response\GetRequisitionResponse;
 use App\Services\Nordigen\Response\NewRequisitionResponse;
 use App\Services\Nordigen\TokenManager;
@@ -82,6 +84,7 @@ class LinkController extends Controller
             $accessToken = TokenManager::getAccessToken();
             $reference   = array_shift($requisitions);
             $request     = new GetRequisitionRequest($url, $accessToken, $reference);
+            $request->setTimeOut(config('importer.connection.timeout'));
             /** @var GetRequisitionResponse $result */
             $result = $request->get();
 
@@ -94,10 +97,19 @@ class LinkController extends Controller
         $uuid        = (string) Uuid::uuid4()->toString();
         $url         = config('nordigen.url');
         $accessToken = TokenManager::getAccessToken();
+
+        $agreementRequest = new PostNewUserAgreement($url, $accessToken);
+        $agreementRequest->setTimeOut(config('importer.connection.timeout'));
+        $agreementRequest->setBank($configuration->getNordigenBank());
+        $agreementRequest->setAccessValidForDays("90");
+        $agreementRequest->setMaxHistoricalDays($configuration->getNordigenMaxDays());
+        $agreementResponse = $agreementRequest->post();
+
         $request     = new PostNewRequisitionRequest($url, $accessToken);
         $request->setTimeOut(config('importer.connection.timeout'));
         $request->setBank($configuration->getNordigenBank());
         $request->setReference($uuid);
+        $request->setAgreement($agreementResponse->id);
 
         app('log')->debug(sprintf('Reference is "%s"', $uuid));
 
@@ -113,7 +125,6 @@ class LinkController extends Controller
         session()->put(Constants::CONFIGURATION, $configuration->toArray());
 
         return redirect($response->link);
-
     }
 
     /**
@@ -130,7 +141,7 @@ class LinkController extends Controller
         app('log')->debug(sprintf('Now at %s', __METHOD__));
         app('log')->debug(sprintf('Reference is "%s"', $reference));
 
-        if('' === $reference) {
+        if ('' === $reference) {
             throw new ImporterHttpException('The reference returned by Nordigen was unexpectedly empty.');
         }